pyproject.toml
[project]
name = "sst0086-frauddetector"
version = "0.1.0"
description = "Add your description here"
license = "MIT"
readme = "README.md"
authors = [
    { name = "Andrea Malloni", email = "malloni.andrea02@gmail.com" },
    { name = "Daniele Monaldi", email = "daniele-monaldi@hotmail.it" },
    { name = "Alessio Rubicini", email = "alessiorubicini05@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "jsonschema>=4.23.0",
    "kagglehub>=0.3.11",
    "lime>=0.2.0.1",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pyspark>=3.5.5",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "setuptools>=80.1.0",
    "shap>=0.47.2",
    "typer-slim>=0.15.2",
    "types-pyyaml>=6.0.12.20250402",
    "xgboost[spark]>=2.1.4",
]

[dependency-groups]
nox = [
    "nox",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-randomly",
]
type_check = [
    "mypy",
    # Add "types-" stub packages as needed: https://github.com/python/typeshed/tree/main/stubs
]
lint = [
    "ruff",
]
docs = [
    "mkdocs-material",
    "mkdocs-htmlproofer-plugin",

    # Python API documentation (not neccessary for applications).
    "mkdocstrings[python]",
    # Autodoc.
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
default-groups = "all"

[tool.hatch.build.targets.wheel]
packages = ["src/fraud_detector"]

[project.urls]
Homepage = "https://github.com/AndreaMalloni/SST0086-FraudDetector/"
Documentation = "https://AndreaMalloni.github.io/SST0086-FraudDetector/"
Repository = "https://github.com/AndreaMalloni/SST0086-FraudDetector/"

[tool.mypy]
ignore_missing_imports = true
strict = true
# TODO: Remove this when explicit-override is enabled by default in strict mode
#   https://github.com/python/mypy/issues/17511
enable_error_code = ["explicit-override"]
# If certain strict config options are too pedantic for a project,
# disable them selectively here by setting to false.

[tool.ruff]
line-length = 99
src = ["src"]
# Ruff will automatically exclude all files listed in .gitignore as well as common temporary Python
# tool directories.
# To exclude additional folders, use extend-exclude.

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "RUF", # ruff
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "PT", # flake-pytest-style
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
extend-ignore = [
    "RUF005",
    "RUF012",
]
unfixable = [
    # Disable removing unused imports by default and only enable within nox so editors don't delete
    # unused imports while the user is in the middle of editing a file on save.
    "F401",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false
# For non-src directory projects, explicitly set top level package names:
# known-first-party = ["my-app"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["typer.Argument"]

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
]
xfail_strict = true
filterwarnings = [
    # When running tests, treat warnings as errors (e.g. -Werror).
    # See: https://docs.pytest.org/en/latest/reference/reference.html#confval-filterwarnings
    "error",
    # Add additional warning suppressions as needed here. For example, if a third-party library
    # is throwing a deprecation warning that needs to be fixed upstream:
    # "ignore::DeprecationWarning:typer",
]

[tool.coverage.run]
branch = true
